requirements.txt
altgraph==0.17.3
bottle==0.12.23
bottle-websocket==0.2.9
cachetools==5.2.1
cffi==1.15.1
Eel==0.14.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
pefile==2022.5.30
pycparser==2.21
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
pywin32-ctypes==0.2.0
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
beautifulsoup4
requests
pillow